requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
google-generativeai>=0.3.2
python-dotenv>=1.0.0
httpx>=0.27.0
sqlmodel>=0.0.14
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
bcrypt==4.0.1
argon2-cffi>=23.1.0
